stop loading enterpriser license info repeatedly
